+++ requirements.txt
@@ -1,0 +1,221 @@
+absl-py==1.3.0
+alembic==1.9.3
+anyio==3.6.2
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+arrow==1.2.3
+asgiref==3.6.0
+asttokens==2.0.5
+astunparse==1.6.3
+async-generator==1.10
+attrs==21.4.0
+Automat==20.2.0
+autopep8==1.6.0
+backcall==0.2.0
+bcrypt==4.0.1
+beautifulsoup4==4.10.0
+bleach==6.0.0
+bs4==0.0.1
+cachetools==5.2.0
+certifi==2021.10.8
+cffi==1.15.0
+charset-normalizer==2.0.12
+click==8.1.3
+colorama==0.4.4
+comm==0.1.2
+community==1.0.0b1
+constantly==15.1.0
+cryptography==36.0.2
+cssselect==1.1.0
+cycler==0.11.0
+debugpy==1.6.6
+decorator==5.1.1
+defusedxml==0.7.1
+Django==4.1.7
+entrypoints==0.4
+et-xmlfile==1.1.0
+exceptiongroup==1.1.0
+executing==0.8.3
+fastjsonschema==2.16.3
+filelock==3.6.0
+Flask==2.2.2
+Flask-Migrate==4.0.4
+Flask-SQLAlchemy==3.0.3
+Flask-WTF==1.1.1
+flatbuffers==22.12.6
+fonttools==4.33.3
+fqdn==1.5.1
+gast==0.4.0
+google-auth==2.15.0
+google-auth-oauthlib==0.4.6
+google-pasta==0.2.0
+greenlet==2.0.2
+grpcio==1.51.1
+gunicorn==20.1.0
+h11==0.14.0
+h5py==3.7.0
+hyperlink==21.0.0
+idna==3.3
+imbalanced-learn==0.10.1
+imblearn==0.0
+incremental==21.3.0
+ipykernel==6.21.2
+ipython==8.2.0
+ipython-genutils==0.2.0
+ipywidgets==8.0.4
+isoduration==20.11.0
+itemadapter==0.5.0
+itemloaders==1.0.4
+itsdangerous==2.1.2
+jedi==0.18.1
+jieba==0.42.1
+Jinja2==3.1.2
+jmespath==1.0.0
+joblib==1.2.0
+jsonpointer==2.3
+jsonschema==4.17.3
+jupyter==1.0.0
+jupyter-console==6.6.2
+jupyter-events==0.6.3
+jupyter_client==7.4.9
+jupyter_core==5.2.0
+jupyter_server==2.3.0
+jupyter_server_terminals==0.4.4
+jupyterlab-pygments==0.2.2
+jupyterlab-widgets==3.0.5
+keras==2.11.0
+kiwisolver==1.4.2
+libclang==14.0.6
+lxml==4.8.0
+Mako==1.2.4
+Markdown==3.4.1
+MarkupSafe==2.1.1
+matplotlib==3.5.2
+matplotlib-inline==0.1.3
+mistune==2.0.5
+mlxtend==0.21.0
+mysqlclient==2.1.1
+nbclassic==0.5.2
+nbclient==0.7.2
+nbconvert==7.2.9
+nbformat==5.7.3
+nest-asyncio==1.5.4
+networkx==2.8.7
+notebook==6.5.2
+notebook_shim==0.2.2
+numpy==1.22.3
+oauthlib==3.2.2
+opencv-python==4.7.0.68
+openpyxl==3.0.10
+opt-einsum==3.3.0
+outcome==1.2.0
+packaging==21.3
+pandas==1.4.1
+pandocfilters==1.5.0
+parsel==1.6.0
+parso==0.8.3
+pdfminer==20191125
+pickleshare==0.7.5
+Pillow==9.1.1
+platformdirs==3.0.0
+prettytable==3.3.0
+prometheus-client==0.16.0
+prompt-toolkit==3.0.38
+Protego==0.2.1
+protobuf==3.19.6
+psutil==5.9.0
+pure-eval==0.2.2
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycodestyle==2.8.0
+pycparser==2.21
+pycryptodome==3.15.0
+PyDispatcher==2.0.5
+pyecharts==1.9.1
+Pygments==2.11.2
+pygwalker==0.1.4.3
+PyMySQL==1.0.2
+pyOpenSSL==22.0.0
+pyparsing==3.0.9
+PyQt5==5.15.7
+PyQt5-Qt5==5.15.2
+PyQt5-sip==12.11.0
+pyrsistent==0.19.3
+PySocks==1.7.1
+python-dateutil==2.8.2
+python-docx==0.8.11
+python-json-logger==2.0.7
+python-louvain==0.16
+pytz==2022.1
+pywin32==303
+pywinpty==2.0.10
+PyYAML==6.0
+pyzmq==25.0.0
+qtconsole==5.4.0
+QtPy==2.3.0
+queuelib==1.6.2
+radian==0.6.3
+rchitect==0.3.37
+requests==2.27.1
+requests-file==1.5.1
+requests-oauthlib==1.3.1
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+rsa==4.9
+scikit-learn @ file:///C:/Users/Brace/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0/LocalCache/local-packages/Python310/Scripts/scikit_learn-1.1.1-cp310-cp310-win_amd64.whl
+scipy==1.8.1
+scorecardpy==0.1.9.2
+Scrapy==2.6.1
+seaborn==0.12.1
+selenium==4.7.2
+Send2Trash==1.8.0
+service-identity==21.1.0
+simplejson==3.17.6
+six==1.16.0
+sklearn==0.0
+sniffio==1.3.0
+sortedcontainers==2.4.0
+soupsieve==2.3.1
+SQLAlchemy==2.0.3
+sqlparse==0.4.3
+stack-data==0.2.0
+tensorboard==2.11.0
+tensorboard-data-server==0.6.1
+tensorboard-plugin-wit==1.8.1
+tensorflow==2.11.0
+tensorflow-estimator==2.11.0
+tensorflow-intel==2.11.0
+tensorflow-io-gcs-filesystem==0.29.0
+termcolor==2.2.0
+terminado==0.17.1
+threadpoolctl==3.1.0
+tinycss2==1.2.1
+tldextract==3.2.1
+toml==0.10.2
+tornado==6.2
+tqdm==4.64.1
+traitlets==5.9.0
+trio==0.22.0
+trio-websocket==0.9.2
+Twisted==22.4.0
+twisted-iocpsupport==1.0.2
+typing_extensions==4.5.0
+tzdata==2022.7
+uri-template==1.2.0
+urllib3==1.26.8
+w3lib==1.22.0
+wcwidth==0.2.5
+webcolors==1.12
+webencodings==0.5.1
+websocket-client==1.5.1
+Werkzeug==2.2.2
+widgetsnbextension==4.0.5
+wordcloud @ file:///C:/Users/Brace/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0/LocalCache/local-packages/Python310/Scripts/wordcloud-1.8.1-cp310-cp310-win_amd64.whl
+wrapt==1.14.1
+wsproto==1.2.0
+WTForms==3.0.1
+XlsxWriter==3.0.3
+zope.interface==5.4.0
+
+pathlib~=1.0.1
+fitz~=0.0.1.dev2
